제갈은경/판매처-POS화면완료-매출이력조회중/2025-10-17

+++ server/scm/src/main/java/com/yedam/scm/vo/SalesDetailVO.java
@@ -10,6 +10,6 @@ public class SalesDetailVO {
     private String saleProdName;    // 제품명
     private Integer saleProdPrice;  // 판매단가
     private Integer saleQty;        // 판매수량
-    private Double saleMargin;      // 판매 당시 마진율
+    private Integer saleMargin;      // 판매 당시 마진율
     private Integer saleUnitPrice;  // 실제 판매단가 (마진 적용된 금액)
 }
